+++ Word2Vec.Net/Model/Comparer/VocubComparer.cs
